Moved CompilerOptions struct to Compiler.h to remove requirement for FileSystemUtils.h, vector and string from external interface.
